Flattener tests: Fix historic_date_format_empty_multivalue.input.xml test
JS & Python parse broken XML differently.
In historic_date_format_empty_multivalue.input.xml - the
      <result type="1">
markup is broken as it is never closed.
However JS and Python assume different names for following elements in this case.

Move broken element to end of <iati-activity >
That way we will still test the broken XML parsing,
but the names will match what is in the expected file now.